3lab/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
websockets==12.0
celery==5.3.6
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
websocket-client==1.7.0
click==8.1.7
tqdm==4.66.2 